Classify APIs by call method
